agents/requirements.txt
langchain
langchain-openai
openai
python-dotenv
tiktoken
ipykernel
pydantic 

faiss-cpu 
langchain-community 
tiktoken
